chore : CollectionViewLayout extension private 추가
